prototype/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-mlx90640==1.3.3
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.76.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.8
cmapy==0.6.6
colorzero==2.0
contourpy==1.3.1
cycler==0.12.1
distro==1.8.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.12.0
gpiozero==2.0.1
idna==3.3
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lgpio==0.2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pigpio==1.78
pillow==11.0.0
pithermalcam==1.0.0
psutil==6.1.1
py-cpuinfo==9.0.0
pycairo==1.20.1
pycryptodomex==3.11.0
pyftdi==0.56.0
PyGObject==3.42.2
pyparsing==3.2.0
pyserial==3.5
python-apt==2.6.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyusb==1.3.1
PyYAML==6.0.2
requests==2.28.1
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
sympy==1.13.1
sysv-ipc==1.1.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.52
ultralytics-thop==2.0.13
urllib3==1.26.12
Werkzeug==3.1.3